requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
Pillow
altair<5